pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epycloud"
version = "0.3.1"
description = "CLI for epymodelingsuite cloud pipeline management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "MOBS Lab", email = "contact@mobs-lab.org"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "google-cloud-storage>=2.18.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "epydemix>=1.0.1",
    "dill>=0.4.0",
    "python-json-logger>=2.0.7",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "epymodelingsuite @ git+ssh://git@github.com/mobs-lab/epymodelingsuite.git@dev",
]

[project.scripts]
epycloud = "epycloud.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = []

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
